src/requirements.txt
matplotlib~=3.3.3
pandas~=1.1.4
numpy~=1.19.3
statsmodels~=0.12.0
pillow~=8.0.1
qiskit~=0.22.0
ipython~=7.16.1
pyflux~=0.4.15
setuptools~=50.3.1
seaborn~=0.11.1